Change from 16 bit int to 32 bit int; baud rate too big to fit

+++ include/driftless/config/OrangeConfig.hpp
@@ -379,7 +379,7 @@ class OrangeConfig : public IConfig {
 
   static constexpr double ODOMETRY_SENSOR_LOCAL_THETA_OFFSET{0.0};
 
-  static constexpr uint16_t ODOMETRY_BAUD_RATE{74880};
+  static constexpr int ODOMETRY_BAUD_RATE{74880};
 
   // ring sensor
